requirements.txt
panda3d==1.10.13
panda3d-gltf==1.0.0
pyinstaller==5.13.0
Cython==3.0.0
pyjwt==2.8.0
flask==2.3.2
flask-sqlalchemy==3.0.5